feat: consolidate admin pages
